Refactor: Move JavaScript to separate file for better organization
